refactor: cars 관련 정보 info로 분리
