+++ src/main/java/racing/domain/CarRacing.java
@@ -3,7 +3,9 @@ package racing.domain;
 import common.Verify;
 import org.apache.commons.lang3.StringUtils;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
 
 import static racing.domain.CarRacingProperty.*;
 
@@ -11,13 +13,13 @@ public class CarRacing {
 
     private static final String TRY_COUNT_ERROR_MESSAGE = "경주 시도 횟수는 1 이상이어야 합니다";
     private static final String INVALID_NAMES_ERROR_MESSAGE = "Car names input not allowed blank";
-    private static final String RACE_NOT_COMPLETED_MESSAGE = "아직 레이스가 끝나지 않았습니다";
 
     private final int tryCount;
     private final Random random;
     private final List<Car> cars;
     private boolean complete = false;
     private int racingCount = 0;
+    private CarRacingInfo racingInfo;
 
     public CarRacing(String inputNames, int tryCount) {
 
@@ -30,6 +32,7 @@ public class CarRacing {
         cars = new ArrayList<>();
 
         parseInputNames(inputNames);
+        racingInfo = new CarRacingInfo(this, cars);
     }
 
     public void race() {
@@ -46,30 +49,12 @@ public class CarRacing {
         return complete;
     }
 
-    public int getCarCount() {
-        return cars.size();
-    }
-
     public int getRacingCount() {
         return racingCount;
     }
 
-    public String[] getCarNames() {
-        return cars.stream()
-                .map(Car::getName)
-                .toArray(String[]::new);
-    }
-
-    public String[] getWinners() {
-        Verify.checkState(complete, RACE_NOT_COMPLETED_MESSAGE);
-        return cars.stream()
-                .filter(car -> car.getDistance() == getMaxMovedDistance())
-                .map(Car::getName)
-                .toArray(String[]::new);
-    }
-
-    public int getMaxMovedDistance() {
-        return Collections.max(cars, Comparator.comparingInt(Car::getDistance)).getDistance();
+    public CarRacingInfo getRacingInfo() {
+        return racingInfo;
     }
 
     public int getRandomDistance() {
@@ -91,5 +76,6 @@ public class CarRacing {
             cars.add(new Car(name));
         }
     }
+
 
 }
